requirements.txt
anthropic>=0.40.0
openai>=1.50.0
streamlit>=1.40.0
requests>=2.32.0
python-dotenv>=1.0.0
pandas>=2.2.0
